fix: consolidate duplicated getPrefersReducedMotion logic into a shared hook
